017_ai_agents_002/requirements.txt
openai-agents==0.0.12
